Setup welcome views/split off auth action
